refactor: replace dashboard and leaderboard pages with dedicated components for better structure
