+++ src/components/Quiz/Game.tsx
@@ -318,7 +318,7 @@ export default function Game({
             </div>
 
             {/* Quiz Content */}
-            <div className='flex min-h-[calc(100vh-200px)] items-center justify-center px-4 py-2 md:py-12'>
+            <div className='flex min-h-[calc(100vh-200px)] items-center justify-center px-4 py-12'>
                 <motion.div
                     className='relative mx-auto max-w-3xl'
                     initial={{ y: 20, opacity: 0 }}
@@ -364,6 +364,7 @@ export default function Game({
                     </div>
 
                     {/* Options Grid */}
+                    {/* //! SAME OPTIONS ARE NOT ANIMATED  */}
                     <div className='grid grid-cols-1 gap-4 md:grid-cols-2'>
                         {quizQuestions[currentQuestion].options.map(
                             (option, index) => (
